Implementada seguridad de control de acceso

+++ src/Controller/TareaController.php
@@ -10,6 +10,7 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Security\Core\Security;
 
 #[Route('/tarea')]
 class TareaController extends AbstractController
@@ -17,14 +18,19 @@ class TareaController extends AbstractController
     #[Route('/listado', name: 'app_tarea_index', methods: ['GET'])]
     public function index(TareaRepository $tareaRepository): Response
     {
+        // deniego el acceso si no tiene ROL de ADMIN
+        $this->denyAccessUnlessGranted('ROLE_ADMIN');
+
         return $this->render('tarea/index.html.twig', [
             'tareas' => $tareaRepository->findAll(),
         ]);
     }
 
     #[Route('/nueva', name: 'app_tarea_new', methods: ['GET', 'POST'])]
-    public function new(Request $request, TareaRepository $tareaRepository): Response
+    public function new(Security $security, Request $request, TareaRepository $tareaRepository): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_USER');
+
         $tarea = new Tarea();
         $form = $this->createForm(TareaType::class, $tarea);
         
@@ -33,6 +39,9 @@ class TareaController extends AbstractController
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
+            
+            // añado el usuario a la tarea que se cree, de momento no puede ser nulo
+            $tarea->setUsuario($security->getUser());
             $tareaRepository->add($tarea, true);
 
             return $this->redirectToRoute('app_listado_tarea', [], Response::HTTP_SEE_OTHER);
@@ -47,6 +56,8 @@ class TareaController extends AbstractController
     #[Route('/{id}', name: 'app_tarea_show', methods: ['GET'])]
     public function show(Tarea $tarea): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_USER');
+
         return $this->render('tarea/show.html.twig', [
             'tarea' => $tarea,
         ]);
@@ -55,6 +66,8 @@ class TareaController extends AbstractController
     #[Route('/{id}/editar', name: 'app_tarea_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Tarea $tarea, TareaRepository $tareaRepository): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_USER');
+
         $form = $this->createForm(TareaType::class, $tarea);
         $form->handleRequest($request);
 
@@ -73,6 +86,8 @@ class TareaController extends AbstractController
     #[Route('/{id}', name: 'app_tarea_delete', methods: ['POST'])]
     public function delete(Request $request, Tarea $tarea, TareaRepository $tareaRepository): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_USER');
+
         if ($this->isCsrfTokenValid('delete'.$tarea->getId(), $request->request->get('_token'))) {
             $tareaRepository->remove($tarea, true);
         }
@@ -83,6 +98,8 @@ class TareaController extends AbstractController
     #[Route('/tarea/{id}', name: 'finalizar_tarea', methods: ['POST'])]
     public function finalizar(Tarea $tarea, Request $request, ManagerRegistry $doctrine): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_USER');
+
         // validamos que la petición sea por AJAX
         if ($request->isXmlHttpRequest()) {
             $em    = $doctrine->getManager();
